data: Move file review fingerprints
File-review safety state persists the selected file identity, page scope,
and action metadata used by pathless follow-up commands.

The module also owns snapshot and diff hashing, so rendering and
suggest-fixup code import persistence state for fingerprint data that is
separate from validation.

This commit addresses that by moving file-review fingerprint construction
into data.file_review_fingerprints and updating state, rendering, and
suggest-fixup callers to import the narrower boundary.

The next commit will add an architecture guard that keeps fingerprint
callers on that boundary.
